Feature/validation logic

+++ assets/js/weightConverter.js
@@ -1,23 +1,23 @@
 export function lbToKilo (inputValue) {
-    return parseInt(inputValue) * 0.4535;
+    return parseFloat(inputValue) * 0.4535;
 }
 
 export function lbToTon  (inputValue) {
-    return parseInt(inputValue) * 0.0004;
+    return parseFloat(inputValue) * 0.0004;
 }
 
 export function kgToPound (inputValue) {
-    return parseInt(inputValue) * 2.2046;
+    return parseFloat(inputValue) * 2.2046;
 }
 
 export function kgToTon (inputValue) {
-    return parseInt(inputValue) / 1000;
+    return parseFloat(inputValue) / 1000;
 }
 
 export function tToKilo (inputValue) {
-    return parseInt(inputValue) * 1000;
+    return parseFloat(inputValue) * 1000;
 }
 
 export function tToPound (inputValue) {
-    return parseInt(inputValue) * 2204.6226;
+    return parseFloat(inputValue) * 2204.6226;
 }
